# perf(es/minifier): Make passes more parallel